Modularize Fader component
