+++ src/components/Fader.scss
@@ -1,0 +1,4 @@
+.fader {
+  opacity: 0;
+  pointer-events: none;
+}
